+++ docs/api/common/multiscales.md
@@ -1,0 +1,3 @@
+# Multiscales
+
+::: ome_zarr_models.common.multiscales
