+++ mkdocs.yml
@@ -34,7 +34,7 @@ markdown_extensions:
 
 plugins:
   - mkdocs-jupyter:
-      execute: true
+      execute: false
       allow_errors: false
   - mkdocstrings:
       handlers:
@@ -70,6 +70,8 @@ nav:
           - api/v04/image-label.md
           - api/v04/labels.md
           - api/v04/well.md
+      - common:
+          - api/common/multiscales.md
   - Changelog: changelog.md
   - Contributing: contributing.md
