+++ src/ome_zarr_models/common/__init__.py
@@ -1,0 +1,3 @@
+"""
+Common functionality between versions of the specification.
+"""
